test: extract shared multisig test helpers and review cleanups
