+++ src/renderer/features/multisig-operations/model/reject-model.test.ts
@@ -1,77 +1,29 @@
-import { type ApiPromise } from '@polkadot/api';
 import { BN } from '@polkadot/util';
 import { type Scope, allSettled, fork } from 'effector';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import {
-  type Chain,
-  type ChainId,
-  type MultisigAccount,
-  AccountType,
-  AssetType,
-  CryptoType,
-  SigningType,
-  TransactionType,
-} from '@/shared/core';
+import { type MultisigAccount, TransactionType } from '@/shared/core';
 import { type AccountId } from '@/shared/polkadotjs-schemas';
 import {
   type AnyAccount,
   type MultisigOperation,
   MultisigEventStatus,
-  accountService,
   accounts,
   transactionService,
 } from '@/domains/network';
 import { networkModel } from '@/entities/network';
 
 import { rejectModel } from './reject-model';
-
-const TEST_CHAIN_ID = '0x0000000000000000000000000000000000000000000000000000000000000001' as ChainId;
-
-const testChain = {
-  chainId: TEST_CHAIN_ID,
-  name: 'Test chain',
-  addressPrefix: 0,
-  options: [],
-  assets: [{ assetId: 0, symbol: 'TST', name: 'Test', precision: 10, type: AssetType.NATIVE }],
-  nodes: [],
-} as unknown as Chain;
-
-const aId = (index: number): AccountId => `0x${String(index).padStart(64, '0')}` as AccountId;
-
-const vaultAccount = (accountId: AccountId, walletId = 1): AnyAccount =>
-  ({
-    id: `account-${accountId}-${walletId}`,
-    walletId,
-    accountId,
-    name: `signer-${walletId}`,
-    type: 'universal',
-    accountType: AccountType.BASE,
-    cryptoType: CryptoType.SR25519,
-    signingType: SigningType.POLKADOT_VAULT,
-    createdAt: 0,
-  }) as unknown as AnyAccount;
-
-const watchOnlyAccount = (accountId: AccountId, walletId = 1): AnyAccount =>
-  ({
-    ...(vaultAccount(accountId, walletId) as object),
-    accountType: AccountType.WATCH_ONLY,
-    signingType: SigningType.WATCH_ONLY,
-  }) as unknown as AnyAccount;
-
-const createMultisigAccount = (accountId: AccountId, signatories: AccountId[], threshold = 2): MultisigAccount =>
-  ({
-    id: `multisig-${accountId}`,
-    walletId: 99,
-    accountId,
-    name: 'Test multisig',
-    type: 'universal',
-    accountType: AccountType.MULTISIG,
-    cryptoType: CryptoType.SR25519,
-    signatories: signatories.map(signatoryId => ({ accountId: signatoryId })),
-    threshold,
-    createdAt: 0,
-  }) as unknown as MultisigAccount;
+import {
+  TEST_CHAIN_ID,
+  aId,
+  createFakeApi,
+  createMultisigAccount,
+  seedDiHandlers,
+  testChain,
+  vaultAccount,
+  watchOnlyAccount,
+} from './test-helpers';
 
 type OperationParams = {
   multisigAccountId: AccountId;
@@ -105,47 +57,6 @@ const createOperation = ({ multisigAccountId, depositor, callHash = '0xabc123' }
     ],
     timestamp: 0,
   }) as unknown as MultisigOperation;
-
-const createFakeApi = (): ApiPromise => {
-  const extrinsic = { isFakeExtrinsic: true };
-
-  return {
-    tx: {
-      multisig: {
-        asMulti: Object.assign(
-          vi.fn(() => extrinsic),
-          { meta: { args: [] } },
-        ),
-        approveAsMulti: vi.fn(() => extrinsic),
-        cancelAsMulti: vi.fn(() => extrinsic),
-      },
-    },
-  } as unknown as ApiPromise;
-};
-
-/**
- * DI anyOf handlers resolve to empty scoped values under fork (a forked scope
- * starts from store initial values), so the availability and permission
- * handlers must be registered in-scope. They mirror production semantics:
- * universal accounts are reachable everywhere; watch-only accounts cannot act.
- */
-const seedDiHandlers = async (scope: Scope) => {
-  await allSettled(accountService.accountAvailabilityOnChainAnyOf.registerHandler, {
-    scope,
-    params: {
-      available: () => true,
-      body: ({ account, chain }) =>
-        account.type === 'universal' || ('chainId' in account && account.chainId === chain.chainId),
-    },
-  });
-  await allSettled(accountService.accountActionPermissionAnyOf.registerHandler, {
-    scope,
-    params: {
-      available: () => true,
-      body: ({ account }) => account.signingType !== SigningType.WATCH_ONLY,
-    },
-  });
-};
 
 const createScope = async (accountList: (AnyAccount | MultisigAccount)[]) => {
   const scope = fork({
